+++ src/Encoder/BooleanEncoder.php
@@ -5,7 +5,7 @@ namespace Riimu\Kit\PHPEncoder\Encoder;
 /**
  * Encoder for boolean values.
  * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
- * @copyright Copyright (c) 2014-2017 Riikka Kalliomäki
+ * @copyright Copyright (c) 2014-2018 Riikka Kalliomäki
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class BooleanEncoder implements Encoder
@@ -22,7 +22,7 @@ class BooleanEncoder implements Encoder
 
     public function supports($value)
     {
-        return is_bool($value);
+        return \is_bool($value);
     }
 
     public function encode($value, $depth, array $options, callable $encode)