+++ src/Encoder/GMPEncoder.php
@@ -5,7 +5,7 @@ namespace Riimu\Kit\PHPEncoder\Encoder;
 /**
  * Encoder for GMP number types.
  * @author Riikka Kalliomäki <riikka.kalliomaki@gmail.com>
- * @copyright Copyright (c) 2014-2017 Riikka Kalliomäki
+ * @copyright Copyright (c) 2014-2018 Riikka Kalliomäki
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class GMPEncoder implements Encoder
@@ -17,7 +17,7 @@ class GMPEncoder implements Encoder
 
     public function supports($value)
     {
-        return is_object($value) && get_class($value) === \GMP::class;
+        return \is_object($value) && \get_class($value) === \GMP::class;
     }
 
     public function encode($value, $depth, array $options, callable $encode)